+++ servant-multipart/servant-multipart.cabal
@@ -28,6 +28,7 @@ library
   default-language: Haskell2010
   hs-source-dirs:   src
   exposed-modules:  Servant.Multipart
+                    Servant.Multipart.Internal
 
   -- ghc boot libs
   build-depends:
