+++ servant-multipart/src/Servant/Multipart/Internal.hs
@@ -1,0 +1,29 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE TypeOperators #-}
+
+module Servant.Multipart.Internal (LookupContext(..)) where
+
+import Data.Proxy (Proxy)
+import Servant.Server (Context((:.)))
+
+-- Utility class that's like HasContextEntry
+-- but allows the lookup to fail, to make a context
+-- entry for upload config optional (hence using
+-- some default configuration when missing)
+class LookupContext ctx a where
+  lookupContext :: Proxy a -> Context ctx -> Maybe a
+
+instance LookupContext '[] a where
+  lookupContext _ _ = Nothing
+
+instance {-# OVERLAPPABLE #-}
+         LookupContext cs a => LookupContext (c ': cs) a where
+  lookupContext p (_ :. cxts) =
+    lookupContext p cxts
+
+instance {-# OVERLAPPING #-}
+         LookupContext cs a => LookupContext (a ': cs) a where
+  lookupContext _ (c :. _) = Just c
